refactor: move event handlers to core root and remove redundant toDouble() call

+++ src/main/java/com/mapconductor/mapbox/marker/MapboxMarkerController.kt
@@ -98,7 +98,7 @@ class MapboxMarkerController private constructor(
 
         val icon = nearest.state.icon ?: DefaultMarkerIcon()
 
-        val baseSizePx = ResourceProvider.dpToPxForBitmap(icon.iconSize).toDouble()
+        val baseSizePx = ResourceProvider.dpToPxForBitmap(icon.iconSize)
         val iconWidthPx = baseSizePx * icon.scale.toDouble()
         val iconHeightPx = baseSizePx * icon.scale.toDouble()
